✅[CHORE] ExpandedStateView 코드 분리(#81)
